backend/api/src/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
neo4j==5.14.1
python-dotenv==1.0.1
openai==1.3.7
httpx==0.24.1 